# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EVA API",
        "contact": {
            "email": "gurkalov.dev@gmail.com"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/v1/positions": {
            "get": {
                "tags": [
                    "Position"
                ],
                "description": "Returns all positions",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "maximum number of results to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Position response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Position"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Position"
                ],
                "description": "Creates a new position.",
                "operationId": "createPosition",
                "requestBody": {
                    "description": "Position to add.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewPosition"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Position response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Position"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/positions/{id}": {
            "get": {
                "tags": [
                    "Position"
                ],
                "description": "Returns a position based on a single ID",
                "operationId": "getPosition",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of position to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Position response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Position"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Position"
                ],
                "description": "Update a position based on a single ID.",
                "operationId": "updatePosition",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of position to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Position to update.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewPosition"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Position response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Position"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Position"
                ],
                "description": "Deletes a single position based on the ID.",
                "operationId": "deletePosition",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of position to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Position deleted"
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorModel": {
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "NewPosition": {
                "required": [
                    "name"
                ],
                "properties": {
                    "positions": {
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object"
            },
            "Position": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/NewPosition"
                    },
                    {
                        "required": [
                            "id"
                        ],
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "user_id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "routers": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        }
    }
}